Fix for 32 bits overflow of array sizes
Ignore-this: e6d7511837a0707fb63473e1f3e4cd20

darcs-hash:9aabd90cde2caad24e763b250b9ba819304b218a

+++ src/common.h
@@ -163,7 +163,7 @@ typedef struct _Point Point;
 static Point last_point;
 
 double  * _constant = NULL;
-extern int datasize;
+extern size_t datasize;
 
 #include "grid/boundaries.h"
